+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # CHANGELOG
 
+## v1.5.2
+
+- 📦 Update to Remix v2.8.0
+- 📦 Update to latest packages
+- 📦 Update to latest React canary
+
+## v1.5.1
+
+- 📦 Update to Remix v2.7.2
+- 🔨 Use new `<Layout>` export
+
+## v1.5.0
+
+- 📦 Update to Remix v2.7.1 (first Vite stable support)
+- 📦 Update to latest packages including React canary
+- 🔨 Update from unstable_vitePlugin
+- ✨ Add SVG sprite icons
+- ✨ Add tailwind helper functions
+- ✨ Add Icon component that supports SVG sprite
+- ✨ Add default error boundary and helper functions
+
+## v1.4.4
+
+- 🔨 Update import type annotation in remix.config.js (#5)
+
 ## v1.4.3
 
 - 📦 Update to Remix v2.6.0